Feature/create add location endpoint and test

+++ src/server/utils/files/cleanUploads.ts
@@ -1,10 +1,14 @@
 import fs from "fs/promises";
-import getUploadPath from "../getUploadPath/getUploadPath";
+import getUploadPath from "../getUploadPath/getUploadPath.js";
 
 const cleanUploads = async () => {
   const uploads = await fs.readdir("uploads");
 
   uploads.forEach(async (file) => {
+    if (/mockImage.jpg/.exec(file)) {
+      return;
+    }
+
     await fs.unlink(getUploadPath(file));
   });
 };
